LPS-94707 frontend-js-web - Move to util root, as there is only one crop-related utility
