+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/global.es.js
@@ -9,7 +9,7 @@ import {
 } from './component.es';
 import escape from 'lodash.escape';
 import getCountries from './util/address/get_countries.es';
-import getCropRegion from './util/cropregion/get_crop_region.es';
+import getCropRegion from './util/get_crop_region.es';
 import getRegions from './util/address/get_regions.es';
 import getFormElement from './util/form/get_form_element.es';
 import groupBy from 'lodash.groupby';
